:sparkles: [FolderableManager] Refacto
