+++ src/Entity/Interface/FolderableEntityInterface.php
@@ -2,7 +2,11 @@
 
 namespace App\Entity\Interface;
 
+use App\Entity\Dossier;
+
 interface FolderableEntityInterface
 {
     public function hasParentFolder(): bool;
+
+    public function move(?Dossier $parentFolder): void;
 }
